requirements.txt
Django==1.5.1
distribute==0.6.27
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==17.5
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
boto
django-storages
django-inplaceedit
django-gravatar
django-registration
django-modeltranslation
django-wysiwyg
django-tinymce
python-mimeparse
dateutil
django-tastypie
